requirements-native.txt
pystray>=0.19.5
Pillow>=10.4.0
PySide6>=6.7.2
